example-application/procs/coons-os/requirements.txt
boto3
python-zaqarclient
requests
